# openapi/openapi.yaml
openapi: 3.0.3

info:
  title: MyService API
  description: API для работы с пользователями
  version: 1.0.0

servers:
  - url: /MyService
    description: Базовый путь

paths:
  /user:
    post:
      summary: Добавить пользователя
      tags:
        - user
      operationId: InsertUser

      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'

      responses:
        '200':
          description: Пользователь успешно создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'

        '400':
          description: Некорректный запрос
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                code: 400
                message: "Некорректный запрос"

        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                code: 500
                message: "Внутренняя ошибка сервера"

  /user/{id}:
    get:
      summary: Получить пользователя по ID
      tags:
        - user
      operationId: GetUserById

      parameters:
        - name: id
          in: path
          required: true
          description: ID пользователя
          schema:
            type: integer
            minimum: 1
            example: 1

      responses:
        '200':
          description: Пользователь найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'

        '404':
          description: Пользователь не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                code: 404
                message: "Пользователь не найден"

        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                code: 500
                message: "Внутренняя ошибка сервера"

    put:
      summary: Обновить пользователя
      tags:
        - user
      operationId: UpdateUser

      parameters:
        - name: id
          in: path
          required: true
          description: ID пользователя
          schema:
            type: integer
            minimum: 1
            example: 1

      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'

      responses:
        '200':
          description: Пользователь успешно обновлён
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'

        '400':
          description: Некорректный запрос
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                code: 400
                message: "Некорректный запрос"

        '404':
          description: Пользователь не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                code: 404
                message: "Пользователь не найден"

        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                code: 500
                message: "Внутренняя ошибка сервера"

    delete:
      summary: Удалить пользователя
      tags:
        - user
      operationId: DeleteUser

      parameters:
        - name: id
          in: path
          required: true
          description: ID пользователя
          schema:
            type: integer
            minimum: 1
            example: 1

      responses:
        '204':
          description: Пользователь успешно удалён

        '404':
          description: Пользователь не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                code: 404
                message: "Пользователь не найден"

        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                code: 500
                message: "Внутренняя ошибка сервера"

components:
  schemas:
    User:
      type: object
      required:
        - id
        - name
        - createdAt
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: John Doe
        createdAt:
          type: string
          format: date-time
          example: "2026-05-21T12:34:56Z"

    CreateUserRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
          example: John Doe

    UpdateUserRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
          example: Updated Name

    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
        message:
          type: string